requirements.txt
falcon==2.0.0
uwsgi==2.0.18
gevent==1.4.0
gunicorn==19.9.0
pyyaml==5.1.2
requests
psycopg2-binary